Move common test methods into separate file
